:construction: Final modifications
Unfortunately, I could not find a solution to cancel the long polling connection with NetHTTP for mantain retro-compatibility. NetHTTP class was simply incomplete from product version 26 to 27 and that was a shock to me. If you are using a product version lower or equal than 26 (10.3) you'll not be able to unsubscribe from a topic unless you close your application or service.